requirements.txt
# Core
python-dateutil>=2.8.2
pytz>=2021.3
pandas>=1.3.0
numpy>=1.21.0
scipy>=1.7.0

# Data Processing
pandas-ta>=0.3.14b0
pyarrow>=6.0.0
pandas-market-calendars>=3.2

# Backtesting
backtrader>=1.9.76.123
backtesting>=0.3.3
vectorbt>=0.22.0

# API Clients
polygon-api-client>=1.13.3
requests>=2.26.0
websockets>=10.0
ccxt>=1.60.0

# Database
sqlalchemy>=1.4.23
alembic>=1.7.3
psycopg2-binary>=2.9.1
redis>=4.0.2

# Monitoring & Logging
prometheus-client>=0.13.1
sentry-sdk>=1.4.3
structlog>=21.1.0
python-json-logger>=2.0.2

# Web & API
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Configuration
python-dotenv>=0.19.0
pyyaml>=6.0
toml>=0.10.2

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.16.0
pytest-xdist>=2.5.0

# Utilities
tqdm>=4.62.0
click>=8.0.3
rich>=10.12.0
loguru>=0.6.0

# Data Visualization
matplotlib>=3.4.3
seaborn>=0.11.2
plotly>=5.3.1
mplfinance>=0.12.7b0

# Machine Learning (optional)
scikit-learn>=1.0.0
tensorflow>=2.7.0
torch>=1.10.0

# Documentation
sphinx>=4.2.0
sphinx-rtd-theme>=1.0.0
myst-parser>=0.15.2
